Move getProductById logic to data source
